+++ src/rational/small/ops/with_int.rs
@@ -19,7 +19,6 @@ macro_rules! forwards {
         impl Add<$ty> for $large {
             type Output = Self;
 
-            #[must_use]
             #[inline]
             fn add(mut self, rhs: $ty) -> Self::Output {
                 AddAssign::add_assign(&mut self, rhs);
@@ -30,7 +29,6 @@ macro_rules! forwards {
         impl Add<&$ty> for $large {
             type Output = Self;
 
-            #[must_use]
             #[inline]
             fn add(mut self, rhs: &$ty) -> Self::Output {
                 AddAssign::add_assign(&mut self, rhs);
@@ -41,7 +39,6 @@ macro_rules! forwards {
         impl Add<$ty> for &$large {
             type Output = $large;
 
-            #[must_use]
             #[inline]
             fn add(self, rhs: $ty) -> Self::Output {
                 Add::add(self.clone(), rhs)
@@ -51,7 +48,6 @@ macro_rules! forwards {
         impl Add<&$ty> for &$large {
             type Output = $large;
 
-            #[must_use]
             #[inline]
             fn add(self, rhs: &$ty) -> Self::Output {
                 Add::add(self, *rhs)
@@ -68,7 +64,6 @@ macro_rules! forwards {
         impl Sub<$ty> for $large {
             type Output = Self;
 
-            #[must_use]
             #[inline]
             fn sub(mut self, rhs: $ty) -> Self::Output {
                 SubAssign::sub_assign(&mut self, rhs);
@@ -79,7 +74,6 @@ macro_rules! forwards {
         impl Sub<&$ty> for $large {
             type Output = Self;
 
-            #[must_use]
             #[inline]
             fn sub(mut self, rhs: &$ty) -> Self::Output {
                 SubAssign::sub_assign(&mut self, rhs);
@@ -90,7 +84,6 @@ macro_rules! forwards {
         impl Sub<$ty> for &$large {
             type Output = $large;
 
-            #[must_use]
             #[inline]
             fn sub(self, rhs: $ty) -> Self::Output {
                 Sub::sub(self.clone(), rhs)
@@ -100,7 +93,6 @@ macro_rules! forwards {
         impl Sub<&$ty> for &$large {
             type Output = $large;
 
-            #[must_use]
             #[inline]
             fn sub(self, rhs: &$ty) -> Self::Output {
                 Sub::sub(self, *rhs)
@@ -117,7 +109,6 @@ macro_rules! forwards {
         impl Mul<$ty> for $large {
             type Output = Self;
 
-            #[must_use]
             #[inline]
             fn mul(mut self, rhs: $ty) -> Self::Output {
                 MulAssign::mul_assign(&mut self, rhs);
@@ -128,7 +119,6 @@ macro_rules! forwards {
         impl Mul<&$ty> for $large {
             type Output = Self;
 
-            #[must_use]
             #[inline]
             fn mul(mut self, rhs: &$ty) -> Self::Output {
                 MulAssign::mul_assign(&mut self, rhs);
@@ -139,7 +129,6 @@ macro_rules! forwards {
         impl Mul<$ty> for &$large {
             type Output = $large;
 
-            #[must_use]
             #[inline]
             fn mul(self, rhs: $ty) -> Self::Output {
                 Mul::mul(self.clone(), rhs)
@@ -149,7 +138,6 @@ macro_rules! forwards {
         impl Mul<&$ty> for &$large {
             type Output = $large;
 
-            #[must_use]
             #[inline]
             fn mul(self, rhs: &$ty) -> Self::Output {
                 Mul::mul(self, *rhs)
@@ -166,7 +154,6 @@ macro_rules! forwards {
         impl Div<$ty> for $large {
             type Output = Self;
 
-            #[must_use]
             #[inline]
             fn div(mut self, rhs: $ty) -> Self::Output {
                 DivAssign::div_assign(&mut self, rhs);
@@ -177,7 +164,6 @@ macro_rules! forwards {
         impl Div<&$ty> for $large {
             type Output = Self;
 
-            #[must_use]
             #[inline]
             fn div(mut self, rhs: &$ty) -> Self::Output {
                 DivAssign::div_assign(&mut self, rhs);
@@ -188,7 +174,6 @@ macro_rules! forwards {
         impl Div<$ty> for &$large {
             type Output = $large;
 
-            #[must_use]
             #[inline]
             fn div(self, rhs: $ty) -> Self::Output {
                 Div::div(self.clone(), rhs)
@@ -198,7 +183,6 @@ macro_rules! forwards {
         impl Div<&$ty> for &$large {
             type Output = $large;
 
-            #[must_use]
             #[inline]
             fn div(self, rhs: &$ty) -> Self::Output {
                 Div::div(self, *rhs)
@@ -617,7 +601,7 @@ mod test {
         assert_eq!(R64!(2, 3) + 2, R64!(8, 3));
         assert_eq!(R64!(5, 6) + 7, R64!(7 * 6 + 5, 6));
         assert_eq!(R64!(5, 6) - 7, R64!(-7 * 6 + 5, 6));
-        assert_eq!(R64!(5, 6) + (-7) as i32, R64!(-7 * 6 + 5, 6));
+        assert_eq!(R64!(5, 6) + -7_i32, R64!(-7 * 6 + 5, 6));
         assert_eq!(R64!(-5, 6) + 7, R64!(7 * 6 - 5, 6));
         assert_eq!(R64!(-5, 6) + (-7), -R64!(7 * 6 + 5, 6));
         assert_eq!(R64!(-2, 3) + 2, R64!(4, 3));
